pyproject.toml
[project]
name = "enterprise-baseline-purple"
version = "1.0.0"
description = "Baseline purple agent for Enterprise MCP tasks"
requires-python = ">=3.11,<3.14"

dependencies = [
    "a2a-sdk>=0.1.0",
    "uvicorn>=0.27.0",
    "starlette>=0.36.0",
    "httpx>=0.26.0",
    "langchain>=0.1.0",
    "langchain-openai>=0.0.5",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0,<3.0.0",
]

[project.optional-dependencies]
test = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
